Simplified `uv sync` steps in "build_publish" and "tests" workflows
Now that the `dev` dependency group has been added, there should be no need to explicitly mention the `test` and `build` groups when running `uv sync`. Both of these groups are included in the `dev` group, and `uv` treats the `dev` group as a special case which is included by default when `uv sync` runs.